requirements.txt
aiosqlite==0.21.0
anyio==4.7.0
certifi==2025.8.3
click==8.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.7
pip==25.2
python-dotenv==1.1.0
python-telegram-bot==22.3
setuptools==78.1.1
sniffio==1.3.0
typing_extensions==4.15.0
wheel==0.45.1
